da code xong process va respond, chua test, mai qua can cu test, h di ngu

+++ 2_Workspace/wishock/device.cpp
@@ -66,6 +66,6 @@ void ledToggle(void){
  * @retval      HIGH
  *              LOW
  */
-uint8_t deviceStatus(void){
+int deviceStatus(void){
     return digitalRead(PIN_DEVICE);
 }
